Complete hosted lowering test ownership
- Verify first-use host callback assembly beside hosted lowering.
- Cover value and never implementation tables with exact specialized metadata.
- Keep output models focused on accessors and formatting.
