+++ src/plan/execution/lowering/host/table.rs
@@ -275,3 +275,214 @@ impl LoweringContext {
         (constant_templates, representations, outcome)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::host::{HostFailure, StatelessHostProfile};
+    use crate::plan::execution::function::{BoolFunctionId, IntFunctionId, ValueFunctionEntry};
+    use crate::plan::execution::graph::{BoolLocalId, IntLocalId};
+    use crate::plan::execution::host::{
+        HostFunctionId, HostNeverFunctionId, HostedFunction, HostedFunctionTarget,
+    };
+    use crate::plan::execution::type_::{
+        FunctionType as ExecutionFunctionType, ValueType as ExecutionValueType,
+    };
+    use crate::plan::{FunctionType, ValueType};
+    use crate::{
+        HostModule, HostProviderModule, HostProviderSet, HostedExecution, ModuleSource,
+        PackageSource, compile_typed_host_program, plan_host_program,
+    };
+    use num_bigint::BigInt;
+    use std::convert::Infallible;
+
+    #[test]
+    fn assembles_value_callbacks_in_first_use_order_and_prunes_unused_registrations() {
+        fn positive(value: BigInt) -> bool {
+            value > BigInt::from(0)
+        }
+
+        assert!(positive(BigInt::from(1)));
+        assert!(!positive(BigInt::from(-1)));
+        let math = HostModule::new("host_support", "host/math")
+            .expect("host module should be valid")
+            .with_function("subtract", <BigInt as std::ops::Sub>::sub)
+            .expect("host function should be valid")
+            .with_function("add", <BigInt as std::ops::Add>::add)
+            .expect("host function should be valid")
+            .with_function("unused", <BigInt as std::ops::Add>::add)
+            .expect("host function should be valid")
+            .with_function("positive", positive)
+            .expect("host function should be valid")
+            .with_function("unused_predicate", || false)
+            .expect("host function should be valid");
+        let execution = hosted_execution(
+            r#"
+import host/math
+
+pub fn main() {
+  let added = math.add(1, 2)
+  #(math.subtract(added, 1), math.positive(added))
+}
+"#,
+            vec!["host_support"],
+            HostProviderSet::new([math]).expect("host modules should be unique"),
+        );
+
+        let functions = execution.host_functions.value_functions();
+        assert_eq!(functions.len(), 3);
+        assert!(execution.host_functions.never_functions().is_empty());
+        assert_host_metadata(
+            &functions[0],
+            "host_support",
+            "host/math",
+            "add",
+            FunctionType::new(vec![ValueType::Int, ValueType::Int], ValueType::Int),
+            &[],
+            ExecutionFunctionType::new(
+                vec![ExecutionValueType::Int, ExecutionValueType::Int],
+                ExecutionValueType::Int,
+            ),
+        );
+        assert_host_metadata(
+            &functions[1],
+            "host_support",
+            "host/math",
+            "subtract",
+            FunctionType::new(vec![ValueType::Int, ValueType::Int], ValueType::Int),
+            &[],
+            ExecutionFunctionType::new(
+                vec![ExecutionValueType::Int, ExecutionValueType::Int],
+                ExecutionValueType::Int,
+            ),
+        );
+        assert_host_metadata(
+            &functions[2],
+            "host_support",
+            "host/math",
+            "positive",
+            FunctionType::new(vec![ValueType::Int], ValueType::Bool),
+            &[],
+            ExecutionFunctionType::new(vec![ExecutionValueType::Int], ExecutionValueType::Bool),
+        );
+
+        assert!(matches!(
+            execution.program.functions.int_function(IntFunctionId(0)),
+            ValueFunctionEntry::Host(HostedFunctionTarget::Value(target))
+                if *target == HostFunctionId::new(0, IntLocalId(0))
+        ));
+        assert!(matches!(
+            execution.program.functions.int_function(IntFunctionId(1)),
+            ValueFunctionEntry::Host(HostedFunctionTarget::Value(target))
+                if *target == HostFunctionId::new(1, IntLocalId(0))
+        ));
+        assert!(matches!(
+            execution.program.functions.bool_function(BoolFunctionId(0)),
+            ValueFunctionEntry::Host(HostedFunctionTarget::Value(target))
+                if *target == HostFunctionId::new(2, BoolLocalId(0))
+        ));
+    }
+
+    #[test]
+    fn assembles_never_callbacks_for_each_first_use_specialization() {
+        fn stop(_: BigInt) -> Result<Infallible, HostFailure> {
+            Err(HostFailure::new("stopped"))
+        }
+
+        assert_eq!(stop(BigInt::from(0)), Err(HostFailure::new("stopped")),);
+        let provider = HostProviderModule::<StatelessHostProfile>::new("application", "main")
+            .expect("provider module should be valid")
+            .with_fallible_function("stop", stop)
+            .expect("host function should be valid")
+            .with_fallible_function("unused_stop", stop)
+            .expect("host function should be valid");
+        let providers = HostProviderSet::with_providers(Vec::<HostModule>::new(), [provider])
+            .expect("provider module should be unique");
+        let execution = hosted_execution(
+            r#"
+@external(erlang, "host", "stop")
+fn stop(value: Int) -> result
+
+@external(erlang, "host", "unused_stop")
+fn unused_stop(value: Int) -> result
+
+pub fn main() {
+  let int_stop: fn(Int) -> Int = stop
+  let bool_stop: fn(Int) -> Bool = stop
+  #(int_stop == int_stop, bool_stop == bool_stop)
+}
+"#,
+            Vec::new(),
+            providers,
+        );
+
+        assert!(execution.host_functions.value_functions().is_empty());
+        let functions = execution.host_functions.never_functions();
+        assert_eq!(functions.len(), 2);
+        assert_host_metadata(
+            &functions[0],
+            "application",
+            "main",
+            "stop",
+            FunctionType::new(vec![ValueType::Int], ValueType::Int),
+            &[ValueType::Int],
+            ExecutionFunctionType::new(vec![ExecutionValueType::Int], ExecutionValueType::Int),
+        );
+        assert_host_metadata(
+            &functions[1],
+            "application",
+            "main",
+            "stop",
+            FunctionType::new(vec![ValueType::Int], ValueType::Bool),
+            &[ValueType::Bool],
+            ExecutionFunctionType::new(vec![ExecutionValueType::Int], ExecutionValueType::Bool),
+        );
+
+        assert!(matches!(
+            execution.program.functions.int_function(IntFunctionId(0)),
+            ValueFunctionEntry::Host(HostedFunctionTarget::Never(target))
+                if *target == HostNeverFunctionId::new(0)
+        ));
+        assert!(matches!(
+            execution.program.functions.bool_function(BoolFunctionId(0)),
+            ValueFunctionEntry::Host(HostedFunctionTarget::Never(target))
+                if *target == HostNeverFunctionId::new(1)
+        ));
+    }
+
+    fn assert_host_metadata<Implementation>(
+        function: &HostedFunction<Implementation>,
+        package: &str,
+        module: &str,
+        name: &str,
+        signature: FunctionType,
+        type_arguments: &[ValueType],
+        type_: ExecutionFunctionType,
+    ) {
+        assert_eq!(function.package(), package);
+        assert_eq!(function.module(), module);
+        assert_eq!(function.name(), name);
+        assert_eq!(function.metadata().signature(), &signature);
+        assert_eq!(function.type_arguments(), type_arguments);
+        assert_eq!(function.type_(), &type_);
+    }
+
+    fn hosted_execution(
+        source: &str,
+        dependencies: Vec<&str>,
+        hosts: HostProviderSet<StatelessHostProfile>,
+    ) -> HostedExecution<StatelessHostProfile> {
+        let typed = compile_typed_host_program(
+            "application",
+            "main",
+            [PackageSource::new(
+                "application",
+                dependencies,
+                [ModuleSource::new("main", "main.gleam", source)],
+            )],
+            hosts,
+        )
+        .expect("host source should compile");
+        let plan = plan_host_program(typed).expect("host source should plan");
+        HostedExecution::try_from_module_plan(plan).expect("hosted execution should seal")
+    }
+}
